+++ Libplanet.Tests/Blockchain/Policies/BlockPolicyTest.cs
@@ -138,7 +138,7 @@ namespace Libplanet.Tests.Blockchain.Policies
                 var validAddress = validKey.PublicKey.ToAddress();
                 return tx.Signer.Equals(validAddress)
                     ? null
-                    : new TxPolicyViolationException(tx.Id, "invalid signer");
+                    : new TxPolicyViolationException("invalid signer", tx.Id);
             }
 
             var policy = new BlockPolicy<DumbAction>(validateNextBlockTx: IsSignerValid);
@@ -167,7 +167,7 @@ namespace Libplanet.Tests.Blockchain.Policies
                 var validAddress = validKey.PublicKey.ToAddress();
                 return tx.Signer.Equals(validAddress)
                     ? null
-                    : new TxPolicyViolationException(tx.Id, "invalid signer");
+                    : new TxPolicyViolationException("invalid signer", tx.Id);
             }
 
             //Invalid Transaction with inner-exception
@@ -178,9 +178,9 @@ namespace Libplanet.Tests.Blockchain.Policies
                 return tx.Signer.Equals(validAddress)
                     ? null
                     : new TxPolicyViolationException(
-                        tx.Id,
                         "invalid signer",
-                        new InvalidTxSignatureException(tx.Id, "Invalid Signature"));
+                        tx.Id,
+                        new InvalidTxSignatureException("Invalid Signature", tx.Id));
             }
 
             // Invalid Transaction without Inner-exception