+++ Libplanet/Action/ActionEvaluator.cs
@@ -151,10 +151,15 @@ namespace Libplanet.Action
             currency * 0;
 
         [Pure]
-        internal static FungibleAssetValue? NullTotalSupplyGetter(
-            Currency currency
-        ) =>
-            null;
+        internal static FungibleAssetValue NullTotalSupplyGetter(Currency currency)
+        {
+            if (!currency.TotalSupplyTrackable)
+            {
+                throw TotalSupplyNotTrackableException.WithDefaultMessage(currency);
+            }
+
+            return currency * 0;
+        }
 
         /// <summary>
         /// Retrieves the set of <see cref="Address"/>es that will be updated when
@@ -344,7 +349,7 @@ namespace Libplanet.Action
                             "rehearsal mode.\n" +
                             "See also this exception's InnerException property.";
                         exc = new UnexpectedlyTerminatedActionException(
-                            null, null, null, null, action, message, e);
+                            message, null, null, null, null, action, e);
                     }
                     else
                     {
@@ -371,12 +376,12 @@ namespace Libplanet.Action
                         logger?.Error(
                             "{Message}\nInnerException: {ExcMessage}", innerMessage, e.Message);
                         exc = new UnexpectedlyTerminatedActionException(
+                            innerMessage,
                             preEvaluationHash,
                             blockIndex,
                             txid,
                             stateRootHash,
                             action,
-                            innerMessage,
                             e);
                     }
                 }
@@ -496,7 +501,7 @@ namespace Libplanet.Action
                     block.ProtocolVersion,
                     delta.GetStates,
                     delta.GetBalance,
-                    delta.GetTotalSupplyImpl,
+                    delta.GetTotalSupply,
                     tx.Signer);
 
                 DateTimeOffset startTime = DateTimeOffset.Now;